### 2019-08-09 ProvTools/Rclean#111 @MKLau
Fix CodeDepends install from devtools
See editor comments from ropensci:

https://github.com/ropensci/software-review/issues/327#issuecomment-515526833
