### 2019-08-23 #327 @annakrystalli
Hey @MKLau, sorry for slow reply, I am actually on holiday so will probably be slow getting back to you for the next week also. Apologies!

In the meantime, I found a short term workaround that hopefully we might be able to turn into a more long-term one.

I forked `CodeDepends` and just [added `biocViews:` to the `DESCRIPTION` file](https://github.com/annakrystalli/CodeDepends/blob/master/DESCRIPTION) in my fork: https://github.com/annakrystalli/CodeDepends.

I also changed the remote repository pointer in the `Rclean` `DESCRIPTION` to `annakrystalli/CodeDepends` to install my fork of the package and installation of `RClean` worked! ✨🎉. This setup installs version 1.62 of `graph`. Out of curiosity, what version do you have installed? I quickly ran the tests and they all passed so on the surface this version of `graph` seems to work with `Rclean`.

I'll make a pull request to the `CodeDepends` repo and hope they will merge it. But if not, you could just maintain a fork of `CodeDepends` that your package depends on where you include `biocViews:` in the `DESCRIPTION`. In fact, I would suggest that you make a fork, set it up as I did and use that as your dependency for now. That would allow both myself and reviewers to proceed with installing and reviewing the package. Then, if and once the PR is merged, you can revert back to the upstream `CodeDepends` repo.

Sound good?
