### 2019-09-11 #327 @MKLau
Hi @annakrystalli, no worries! Hope you had a relaxing/restorative holiday.

- **dev and master branch usage**: I had stopped moving master forward as so many changes were happening due to the ROpenSci and JOSS review. Travis is currently running checks on the following updates (see following items). I'll merge into master after those checks pass. However, CRAN will now be behind master. *Do you recommend updating it or waiting for the ROpenSci review process to proceed?*
- **goodpractice Imports**: I had added *BiocManager* to Imports to deal with the installation of graph but I think that your suggested fixes to *CodeDepends* has made that vestigial. *knitr* is required for automatically building the vignette. I removed *BiocManager* entirely and added *knitr* under Suggests.
- **goodpractices spell_check**: fixed all of those, funny how many times I've run spellcheck but typos keep popping up!
- **goodpracices package coverage**: *var.id* has become a legacy function that I think may be useful in the future, so I haven't gotten rid of it, but I also haven't bothered to add unit tests. I've added an issue for that, and I'll add them when I next get a chance. *write.code* only has partial coverage because part of its functionality is copying to the clipboard, which I haven't figured out how to write tests for. If you (or maybe one of the reviewers) have a suggestion, I'd be happy to get the coverage on that function higher.
- **Add badge to README**: done!

Thanks for the updates on this, and happy to see the review move forward.
