### 2019-09-11 #327 @MKLau
@annakrystalli  FYI - master is now up to date with dev (now v1.1.0). CRAN is still v1.0.0.
