### 2019-09-25 #327 @nevrome
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

A very interesting package to simplify R scripts. I really like the underlying idea and the main functions `clean` and `write.code` work well when applied to the example script `example/simple_script.R`. Good job, @MKLau! However, the usability of the package can be increased a bit more. Some suggestions in no particular order:

- The function **documentation** is very brief at the moment. I don't think this necessarily a bad thing, but for some of the functions the lack of description and example code leaves me wondering, what the respective function is supposed to do. Please add some more descriptive text and example code for `clean`, `codeGraph` and `write.code`. I would like to see `codeGraph` mentioned in the README/vignette.
- I would love to have an **example script** (maybe `example/simple_script.R`) as a [package example dataset](http://r-pkgs.had.co.nz/data.html) that allows to explore the functions without the need to provide/prepare an own script. An example script will also allow you to add meaningful example code for the functions.
- **Retrospective Provenance** remained a mystery for me. The missing piece of the puzzle is how to generate the relevant json file. Maybe I just missed the crucial hint? I would prefer to see a more extensive explanation of this feature and how to use it in the README/vignette.
- In the context of the Retrospective Provenance feature you use the **`option` interface** to access the content of the json file. I do not understand why this rather unusual solution is necessary. Why can't the `rd` parameter simply get the relevant character vector? Instead of `options(prov.json = readLines("example/prov_micro.json")); clean(file = "example/micro.R", rp = TRUE)` the interface could simply be `clean(file = "example/micro.R", rp = readLines("example/prov_micro.json"))` or even just the path `clean(file = "example/micro.R", rp = "example/prov_micro.json")`. Why did you decide against this?
- When I tried to apply `clean` to some of my own scripts, I realized that all packages necessary to run this script have to be loaded in the current session. If this was not the case, then I got the error message **`Error in as.environment(pos): no item called "package:dplyr" on the search list`**. I got this message 10+ times for one of my more complex scripts before clean went through, until I loaded all the packages. This is a little bit tedious. Is there another way to implement package loading? Maybe the [automagic](https://github.com/cole-brokamp/automagic) package could help to automate that? If this is not possible or if you don't want to have an automatic solution, then I would at least ask you to catch the current error message and replace it with something more intelligible, e.g. "Package xyz has to be loaded to analyze this script.".
- **Internal functions** do not need to be exported (e.g. `parse.info`, `read.prov`). Remove the `@export` tag for these functions and add the tags `@keywords internal` and `@noRd`.
- The **`simple_script.R`** code appears many times in the package (in the directories `test/`, `inst/`, `exec/`, `example/`). I don't think you need this file so many times.
- Although you not seem to work in RStudio I suggest to add a **Rstudio project file** in the root directory (+ the necessary additions to `.Rbuildignore`). That simplifies contributing for RStudio users.
- You could add a **help page** for the whole package as described [here](http://r-pkgs.had.co.nz/man.html#man-packages). I often use this page as an entry point for the function index.
- This might be way beyond the purpose of this review, but as I personally consider these kind of things when I decide whether I want to use a package, I would like to add that I was not able to access neither your **homepage** http://mklau.info/ nor the one the Github Orga ProvTools http://provtools.org/ [2019-09-24].
- The **JOSS** paper manuscript seems to be in an unfinished state because several sections are mentioned with a keyword but not yet written. The text available so far is fine.
- In the **travis-ci** setup I suggest to treat warnings as errors (`warnings_are_errors: true`). I think this is usual for R packages, because no package with warnings can go to CRAN.
