### 2019-09-27 #327 @wlandau
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

`Rclean` is a promising package. It goes above and beyond typical static code analysis and addresses a huge unmet needs in scientific workflows. Well done, @MKLau!

- Please include a `CODE_OF_CONDUCT.md` file at the top level in your repo. You could generate one [here](https://github.com/ProvTools/Rclean/community) or call `usethis::use_code_of_conduct()`. On [this page](https://github.com/ProvTools/Rclean/community), you can even generate issue and pull request templates that remind people of the code of conduct.

- Consider renaming `CONTRIBUTE.md` to the more widely-used `CONTRIBUTING.md`.

- Do you have a longer, real-world script to show? `simple_script.R` gets the idea across, but if you show how to extract simple chunks from a confusing script with 1000+ lines, you could make `Rclean` really shine.

- At the top of the README, please consider reorganizing those bullet points into two paragraphs or subheadings: one introducing the existing problems in scripts and reproducibility, and then another focusing on how `Rclean` solves those problems. Maybe also elaborate on both points as well. I do not think this is the place to mention the provenance model or graphs, but you could definitely talk more about the survey you mentioned, the problem of messy code in science, and the fact that `Rclean` sifts through code to extract just what you need for a given result.

- Would you cite that survey from the third bullet in the README? If you cannot find the citation, feel free to remove the mention entirely and choose a different way to motivate the problem.

> A recent survey of over 1500 scientists reported a crisis of reproducibility with “selective reporting” being the most cited contributing factor and 80% saying code availability is playing a role.

- You might consider using a `README.Rmd` to generate the `README.md` so it is easier to refresh the docs when you change the package (example [here](https://github.com/tidyverse/dplyr/blob/master/README.Rmd)).

- I strongly agree with @nevrome to add examples to the help files for the exported functions you choose to keep (see below). It is difficult for me to understand how to create arguments for most of your exported functions.

- The [vignette](https://github.com/ProvTools/Rclean/blob/master/vignettes/Rclean.rmd) repeats a lot of information from the README. Let's think about removing the redundant sections and refocusing the vignette on a deeper dive in a feature set not covered in the README. Perhaps retrospective provenance ? See the next comment.

- It is not intuitively clear to me why we need retrospective provenance. In the README or vignette, please demonstrate a situation where prospective provenance is not enough and retrospective provenance gives us what we seek.

- Please consider citing a paper to make it clear that "retrospective" and "prospective" are widely used and accepted terms to describe provenance. A quick search reveals a couple possibilities (examples [here](https://ieeexplore.ieee.org/document/5557202) and [here](https://www.researchgate.net/publication/221212739_Provenance_and_scientific_workflows_Challenges_and_opportunities))?

#### API

- The `clean()` function tries to do three things: (1) list lines of execution, (2) list the names of possible results, and (3) get the code required to produce one of the results. I agree that `clean()` should do (3), but I think (1) and (2) are not really cleaning yet and should go in their own functions. Maybe `list_lines()` (which returns an integer vector) and `list_results()`?

- Similarly, it is also worth considering an entirely new set of functions for retrospective provenance because they have different inputs and [your underlying code is entirely different](https://github.com/ProvTools/Rclean/blob/f6b7ffc3325638ba878d7bb00df28a630043920d/R/clean.R#L86-L182).

- In `help(package = "Rclean")`, I see several API functions that represent very technical steps and esoteric data structures. Are there specific reasons why you exported these things to the user? I am not sure they will want to work with matrix representations of graphs or PROV-JSON-formatted provenance objects in memory. It seems like the easiest inputs to understand are R script files and JSON files. Similarly, the best outputs are human-readable: cleaned-up scripts and plotted code graph visualizations. Maybe even consider writing files directly from `clean()`. Consistency and simplicity are extremely valuable.

- I am questioning the need for most of `Rclean`'s exported functions. Not all your functions need to be exported, and not all of them need `roxygen2` docstrings.
    - Do we really need `p.spine()` and `p.spine()`? If you add a function to get the `igraph` from a script, it seems like you can simply instruct users to call `dfs()` themselves.
    - Should `get.libs()`, `var.lineage()`, and `var.id()` really be exported? `clean()` needs them, but do you expect most users to ever invoke them manually?
    - Do we really need a `write.code()` function? `readLines()` already writes code to a file, and `clipr::write_clip()` writes to the clipboard. Optional: maybe `clean()` itself could have an argument to control where the output goes: either to the clipboard with `write_clip()` or to a file if a path is given.
    - You explicitly list `parse.info()`, `read.prov()`, and `parse.graph()` as internal functions. Perhaps these should not be exported either.

- The names of exported functions have different naming conventions: `codeGraph()` (camel case) vs `prov_json()` (snake case) vs `get.spine()` (dot-separated). For the exported functions you choose to keep, please consider sticking to one of these conventions (I recommend snake case). Depending on the user base you have right now, you may want to gracefully deprecate those functions as [described here](https://ropensci.org/technotes/2017/01/05/package-evolution/).

#### Installation and automated checks

- Re https://github.com/ropensci/software-review/issues/327#issuecomment-520062929, I also had some trouble installing `Rclean`. I needed to install `Rgraphviz` directly from Bioconductor. I suggest adding the following to the installation instructions in the README. At the very least, it is a useful stopgap until @annakrystalli's pull request to `CodeDepends` reaches CRAN.

```r
install.packages("BiocManager")
BiocManager::install(c("graph", "Rgraphviz"))
```

- `devtools::spell_check("Rclean")` identifies several correctly spelled words. I recommend listing them in an `inst/WORDLIST` file ([example here](https://github.com/ropensci/drake/blob/master/inst/WORDLIST)).
- `goddpractice::gp("Rclean")` suggests a bunch of style fixes I agree with. The `styler` package can help you turn assignment `=` into `<-`. Please let me know if you would like clarification on any of these notes.

```r

    R/clean.R:49:NA
    R/clean.R:72:NA
    R/clean.R:82:NA
    R/clean.R:91:NA
    R/clean.R:95:NA

  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/var.id.R:41:44
    R/var.id.R:42:47

    R/clean.R:47:1
    R/clean.R:93:1
    R/get.spine.R:44:1
    R/parse.info.R:39:1
    R/parse.info.R:40:1

    R/clean.R:150:29
    R/clean.R:151:55
    R/clean.R:169:34

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/get.libs.R:40:13
    R/get.libs.R:42:17
    R/get.libs.R:44:21
    R/var.id.R:39:13
    R/var.lineage.R:43:22
```

#### Miscellaneous

- I am a bit concerned about the scope of the built-in unit tests. It would be helpful to test at least one longer script to make sure the correct lines get captured in a serious workflow. Not only that, but when you actually run the longer script and various object-specific scripts generated by `clean()`, do the output data objects agree?

- What is the reason you chose JSON to store retrospective provenance? Have you considered a tidier format such as a long-form data frame? I like how [`pkgapi`](https://github.com/r-lib/pkgapi) represents its output.

- The `clean()` function is ambitious, long, and deeply indented. Please consider decomposing this monolithic function into smaller helper functions. No need to export the helper functions or write `roxygen2` docstrings. A particularly good resource on this kind of refactoring is Jenny Bryan's talk on [code smells and feels](https://www.youtube.com/watch?v=7oyiPBjLAWY).

- Are you open to using [`styler`](https://github.com/r-lib/styler) instead of `formatR` to reformat code? It is more modern and more actively maintained.
