### 2019-10-11 #327 @wlandau
One clarification on a couple points I made:

>  I am questioning the need for most of Rclean's exported functions...Do we really need p.spine() and p.spine()?

Those functions are necessary to implement `clean()`, and I recommend keeping them internally. I only question the need for an `@export` tag because I suspect most users will not invoke these functions directly.

In fact (and this is not strictly necessary to satisfy my review) it could be helpful to split `clean()` up into multiple helper functions. This could reduce the deeply nested logic in your code and make `Rclean` easier to maintain. A couple resources:

- [Jenny Bryan's "Code smells and feels" talk](https://www.youtube.com/watch?v=7oyiPBjLAWY)'s is an excellent and widely cited resource on code style best practices.
- [Cyclomatic complexity](https://en.wikipedia.org/wiki/Cyclomatic_complexity) is a measure of the logical complexity of a program, and the [`cyclocomp` package](https://github.com/MangoTheCat/cyclocomp) reports the cyclomatic complexity of R code chunks and functions.
