### 2019-10-11 #327 @MKLau
Thanks for taking the time to clarify @wlandau. I get what you’re saying both about the structure of the API and the logic of the `clean` function. This review process has been great because it’s forcing me to hone the target user/application, which has meandered a bit in development. I think the inclination to make the spine and other functions internal is the way to go.

Also, thanks for the pointers, especially the cyclocomp package.
