### 2019-10-29 #327 @danielskatz
Please update https://github.com/openjournals/joss-reviews/issues/1312 when this is accepted.
