### 2019-12-06 #327 @wlandau
@MKLau, fantastic work on the revisions! You put a ton of effort into refactoring the code, rethinking the package scope, and strengthening the documentation, and `Rclean` is totally transformed.

After reviewing https://github.com/MKLau/Rclean/commit/b674396e4c84f58ae8a46ac50f7312650c300e07, I have some minor follow-up comments and requests.

- [ ] Would you point me to the new test you added on a longer script? I am not sure I found what you are referring to.
- [ ] Code coverage could be improved if you think it is reasonable. Raising coverage to 100% is not as important as diligently verifying the core claims of the package, but I thought I would raise the issue since `Rclean` seems relatively light in tests. I have included a coverage report below. You could do some combination of adding tests that hit those lines and adding inline comments in the source for lines you want to deliberately exempt.

```r
Rclean Coverage: 89.69%
R/keep.R: 83.33%
R/clean.R: 85.00%
R/get_libs.R: 85.71%
R/get_vars.R: 85.71%
R/code_graph.R: 92.31%
R/var_lineage.R: 100.00%
         filename  functions line value
51      R/clean.R      clean   57     0
53      R/clean.R      clean   58     0
79      R/clean.R   get_path  132     0
81      R/clean.R   get_path  135     0
82      R/clean.R   get_path  136     0
83      R/clean.R   get_path  137     0
20 R/code_graph.R code_graph   46     0
9    R/get_libs.R   get_libs   42     0
18   R/get_vars.R   get_vars   45     0
31       R/keep.R       keep   49     0
```

- [ ] The `goodpractice::gp()` report lists a couple other minor notes too.

```r

    R/code_graph.R:15:1
    R/keep.R:17:1
    R/var_lineage.R:15:1
    R/var_lineage.R:22:1

  ✖ fix this R CMD check NOTE: Namespace in Imports field not
    imported from: ‘jsonlite’ All declared Imports should be used.
```

Without retrospective provenance, it appears you no longer need `jsonlite`. Also, line length is ultimately a matter of personal style, but lines less than 80 characters are relatively easy to read. You could either wrap those lines or let me know what you prefer stylistically.

- [ ] Lastly, I recommend fixing a couple examples. In `package.R`, I recommend using `system.file()` to get the path like you do elsewhere (as opposed to the hard-coded `example/simple_script.R`). And [here](https://github.com/MKLau/Rclean/blob/b674396e4c84f58ae8a46ac50f7312650c300e07/R/keep.R#L39-L43), the value returned by `data()` does not have the lines of code we want. I recommend something like this:

```r
#' data(simple_script)
#' ## Copies code to your clipboard
#' keep(simple_script)
#' ## Saves code to your disk
#' keep(simple_script, file = "clean_code.R")
```
