### 2019-12-09 #327 @nevrome
I agree with @wlandau: You did an excellent job refactoring Rclean, @MKLau. I really enjoyed reading the vignette and running the example code. I will keep this in mind to analyse my own scripts.

Maybe an RStudio Addin would be nice: Select a variable in your code file with your mouse, press a key combo, get the `clean()` result. Would make it even more convinient and is simple enough to implement with `rstudioapi`.

Some last, small remarks:

***

- I agree with @wlandau that the example code should run out of the box without any additional work for the user. That is especially true for the README. I therefore suggest to replace

```
script <- "example/simple_script.R"
```

with

```
script <- system.file("example", "simple_script.R", package = "Rclean")
```

- The plot produced by code_graph has very small node labels. Maybe you could add argument passing with `...` (`code_graph(script, pdf.file, ...)`) to `Rgraphviz::plot.graphNEL` (`plot(gNEL, ...)`). That would allow me to set the label size. I personally would prefer to separate graph generation and graph plotting, but if you combine them, then at least I would like to have some control over the appearance.

- Unrelated to the package: The lack of https access to http://mklau.info/ still causes problems for me in chromium. Maybe your hoster provides an easy way to get Let's encrypt certificates for this domain? It will be increasingly difficult for users to access your homepage without encryption.

- Concerning the JOSS submission: Afaik the JOSS editors assume that the paper is perfectly fine already when it went through the rOpenSci review. Can you confirm that @annakrystalli? The paper is not completely ready yet though, so it does not make sense for me to read and review it at this point. I wonder what's the correct procedure here to make sure that JOSS receives reviewed texts.

***

If you want to add the reviewers to the DESCRIPTION file, then you can use the following string for me:

```
	person(given = "Clemens",
             family = "Schmid",
             role = "rev",
             email = "clemens@nevrome.de",
             comment = c(ORCID = "0000-0003-3448-5715"))
```
