### 2019-12-09 #327 @danielskatz
pinging @labarba as the JOSS editor for that part of this - note the comment about the JOSS paper and review (https://github.com/openjournals/joss-reviews/issues/1312) in the comment above.
