### 2019-12-09 #327 @labarba
When a paper goes through rOpenSci review, it gets handled by @arfon on the JOSS side. I have not handled any of these.
