### 2019-12-09 #327 @danielskatz
Actually, it can be handled by the AEiC on duty - I've done this a few times, as have others.  See https://joss.readthedocs.io/en/latest/editing.html#processing-of-ropensci-reviewed-and-accepted-submissions
