### 2019-12-09 #327 @danielskatz
In any event, to answer @nevrome's question, my understanding is that once this submission is accepted by rOpenSci, the JOSS Associate-Editor-in-Chief responsible for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1312) (which is currently @labarba) will review the paper to make sure it meets JOSS standards, but will assume the review of everything else done by rOpenSci is sufficient.  (But maybe it's worth asking @arfon to make sure I have this right)
