### 2019-12-09 #327 @nevrome
Thank you for your answer, @danielskatz. My issue is mostly caused by the unclear border between code and text review.

@wlandau and I provided code review according to this [Guide for Reviewers](https://devguide.ropensci.org/reviewerguide.html). The guide does not mention the JOSS paper. The [review template](https://devguide.ropensci.org/reviewtemplate.html) contains the following section, though:

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

So obviously it is expected that I read and review the text. Also from my understanding of your answer, @danielskatz, there won't be additional review beyond a rather general check of the editor over at JOSS.

The takeaway message is, that the course of action suggested above by @MKLau is not possible:

> The initial submission of the JOSS paper was based on an example of a submitted article present on the JOSS instructions. However, after this package review, there will be substantial content added to the article. I am currently working on these updates to the text.

If there are significant changes of the article **AFTER** this review, then they won't be reviewed at all.
