### 2019-12-09 #327 @annakrystalli
Thanks @nevrome for picking this up.

What we can do is consider the code review of the package complete and move it to rOpenSci. I will coordinate with @MKLau and when the `paper.md` is complete and ready for JOSS submission, I will review it before it gets submitted (unless of course you want to @nevrome). Ultimately I'm happy to consider your contribution complete and do the final check myself.
