### 2019-12-09 #327 @danielskatz
> If there are significant changes of the article AFTER this review, then they won't be reviewed at all.

Well, they are reviewed by the Associate-Editor-in-Chief, but we would certainly appreciate anything you do that improves the paper. Whether this is an obligation of rOpenSci reviewers of submissions that are also being submitted to JOSS is for rOpenSci to determine.
