### 2019-12-09 #327 @MKLau
Great, thanks to everyone for clarifying the relationship between the software and article review. Thanks for the offer to review the manuscript (@nevrome), it would be great to get your input on it if you have the time to spare.

@annakrystalli, I’ll send you the updated manuscript soon. Hopefully by week’s end.
