### 2019-12-09 #327 @MKLau
@wlandau and @nevrome, thanks also for the follow-up comments and suggestions.

- I'll take care of the NAMESPACE update and the long line issues. I'm not sure why `gp` didn't pick these up when I ran it.
- I've added "enhancement" issues for the RStudio add-in and the node label sizing for code_graph. I'd like to get those implemented soon, but probably after the review, as they're not integral to the functioning of the package.
- @wlandau, per your request for the test reference, I was referring to line 26 in test-main.R. It tests the correct cleaning of code for a two variables that are "tangled" around several other variables. I agree with your general comments about testing. After taking a closer look using `covr` I think should be able to get the coverage up to over 95%, but I agree, it's more important to have tests that are checking for correct functioning of the package. I think that the clean function testing could be more extensive, and if you have more specific suggestions, I'd be happy to integrate them.

Thanks!
