### 2019-12-10 #327 @wlandau
I see. Do you think the simple script is sufficient to test `Rclean`'s ability to sift out "tangled" variables, or are there different kinds of decisions `Rclean` makes that you could explore with the longer script?

Either way, I highly recommend accruing tests for all the bug reports that come your way over time. If you write a new test for each bug as you fix it, your test suite will become powerful over time.

Please ping me when you are ready for me to take another look.
