### 2019-12-10 #327 @MKLau
@wlandau thanks, point well taken. Now that the package is stabilizing (and much improved thanks to this review), I can spend more time thinking about how to test it. For now, it seems to be at a place where it should work enough for folks to use it enough to report errors and not just pass on it and not report anything.
