### 2019-12-11 #327 @annakrystalli
Thanks @MKLau. Just a note that good testing is just as much part of a package maturing as is adding functionality and shouldn't be considered something you do afterward but rather, as you add functionality.

As you say, time and bug reports are important in catching errors from edge cases, but I think what @wlandau (feel free to correct me if I'm misinterpreting!) is trying to get you to do is think defensively about any systematic ways in which different uses of the function could break. This could be different combinations of arguments, different inputs etc. Even 100% coverage does not indicate catching these as it's not about testing a line of code once (which could lead to 100% coverage) but under various realistic use situations.

@wlandau do you have any particular tests you have in mind? Or do you think as @MKLau suggest that the test suite is good enough for this stage?
