### 2019-12-11 #327 @wlandau
@MKLau, I like [these tests](https://github.com/MKLau/Rclean/blob/06c1f681ab6e9e8b3109049bcb27ff347a0b80f9/tests/testthat/test-main.R#L27-L42), and I like that they are specific and targeted. I strongly suggest a couple more tests like them for a couple variables in `long_script.R`.

I also think it is important to compare the actual values of the results you get from actually running the scripts. A cleaned script ultimately needs to produce the same output you get from the original. So for [this test](https://github.com/MKLau/Rclean/blob/06c1f681ab6e9e8b3109049bcb27ff347a0b80f9/tests/testthat/test-main.R#L27-L33), you could

1. Run `source(simple.script, local = some_new_environment)` and inspect `some_new_environment` to get one value of `out`.
2. Similarly, get another value of `out` from the cleaned script.
3. Compare the two values of `out` using `expect_equal()`.

You currently compare the actual lines of code, which I suggest you keep. However (and this is not part of my official review) `rda` files may not be the easiest thing to work with in this case. I think flat `.R` files would be easier to work with. To avoid clutter in the `tests/testthat` directory, you could give each `test-*.R` script its own `test-*/` directory to keep track of expected results.

One idea for a feature (again, not part of my review) is to implement validation as a core feature of `Rclean`: have a function that runs each script in its own fresh R environment or session and asserts that they give identical values for all the variables of interest. [`callr`](https://callr.r-lib.org/) and `tempfile()`s could be handy on the implementation side.

It is easy to fool static code analysis, and I expect unsolvable edge cases to come up eventually. When that happens, I encourage you to document them so users understand how `Rclean`.
