### 2019-12-11 #327 @MKLau
@annakrystalli I definitely agree with the point of defensive testing. These comments of @wlandau’s with more specifics and pointers are great. I’ll get to those test tomorrow and add these suggested features to the issues list as enhancements. Also, I will look into streamlining the bug reporting processes. I have documentation for that but it could be improved and more could be done automatically.

Also, I like the idea of validation. In theory, the code extract should produce the same result, unless there is randomness, etc. If it could be done, it would be great to implement as a part of the RStudio add-in, as suggested previously by @nevrome. That is, when code is extracted for a result(s), it is automatically checked.

Thanks!
