### 2019-12-11 MKLau/Rclean#185 @MKLau
Increase clean testing
From @wlandau:

> @MKLau, I like these tests, and I like that they are specific and targeted. I strongly suggest a couple more tests like them for a couple variables in long_script.R.

> I also think it is important to compare the actual values of the results you get from actually running the scripts. A cleaned script ultimately needs to produce the same output you get from the original. So for this test, you could

> Run source(simple.script, local = some_new_environment) and inspect some_new_environment to get one value of out.

> Similarly, get another value of out from the cleaned script.

> Compare the two values of out using expect_equal().
