### 2019-12-12 #327 @wlandau
> Also, I will look into streamlining the bug reporting processes. I have documentation for that but it could be improved and more could be done automatically.

To help encourage bug reporting, you might have a look at [issue templates](https://help.github.com/en/github/building-a-strong-community/configuring-issue-templates-for-your-repository) like [this one](https://github.com/ropensci/drake/blob/master/.github/ISSUE_TEMPLATE/bug.md).
