### 2019-12-12 #327 @MKLau
@wlandau Just added a new test that checks the reproducibility of variables in cleaned scripts, please see [here](https://github.com/MKLau/Rclean/blob/dev/tests/testthat/test-main.R#L54-L97).

Let me know your thoughts, in particular I'm wondering if you would separate the variables into individual tests.
