### 2019-12-12 MKLau/Rclean#185 @MKLau
Checking for feedback from ROpenSci reviewers.
