### 2019-12-13 MKLau/Rclean#185 @wlandau
Detailed suggestions about https://github.com/MKLau/Rclean/blob/dev/tests/testthat/test-main.R#L54-L97:

1. What about the ability of cleaned scripts to *exclude* variables? You could test that `fit.xx`, `fit.sqrt.A`, and `fit.anova` are in `env.long` but not the other environments.
2. I think it makes sense to keep all the variables in the same test, but maybe you could condense it down to a loop that calls a function. 3. When you get variables from the environments, `get()` with `inherits = FALSE` prevents R from looking in the parent environment for the object, and I think it would add a little safety here.
4. `tempfile()`s are more convenient than creating explicitly named scripts and removing them at the end.
5. `withr::with_seed()` sets the seed for a chunk of code and restores the original seed afterwards. It could be more convenient and maybe safer.

Sketch:

```r
long_script <- system.file(
  "example",
  "long_script.R",
  package = "Rclean"
)
env_long <- new.env()
withr::with_seed(42, source(long_script, local = env_long))
test_clean <- function(var, long_script, env_long) {
  cleaned <- clean(long_script, var)
  tmp <- tempfile()
  keep(cleaned, tmp)
  env_cleaned <- new.env()
  withr::with_seed(42, source(tmp, local = env_cleaned))
  out <- get(var, env_cleaned, inherits = FALSE)
  exp <- get(var, env_long, inherits = FALSE)
  expect_equal(out, exp)
  invisible()
}
vars <- list("x", "x2")
tmp <- lapply(
  vars,
  test_clean,
  long_script = long_script,
  env_long = env_long
)
```
