### 2019-12-13 #327 @wlandau
Nice, that's a great test. I added some suggestions at https://github.com/MKLau/Rclean/issues/185#issuecomment-565273312. In the interest of avoiding a rabbit hole, only (1) is part of my official review, e.g.

> What about the ability of cleaned scripts to exclude variables? You could test that fit.xx, fit.sqrt.A, and fit.anova are in env.long but not the other environments.
