### 2019-12-13 #327 @MKLau
@wlandau I've added two new tests (see [here](https://github.com/MKLau/Rclean/blob/1b565a74e6c21ed02d28f8a72f5542b04fef716a/tests/testthat/test-main.R#L100-L148) that should address your last request in the comment above. They don't directly test for the "exclusion" of variables, instead they test that a set of expected variables are present in the cleaned scripts when sourced in a new environment. I've manually inspected these to verify that they are correct. Let me know if this deviates from the intent of your recommended test.
