### 2019-12-14 #327 @wlandau
I reviewed those new tests, and I think the test suite is now diligent and defensive enough. My follow-up requests are minor:

- [ ] Use `system.file()` in the examples [here](https://github.com/MKLau/Rclean/blob/b674396e4c84f58ae8a46ac50f7312650c300e07/R/keep.R#L39-L43) and [here](https://github.com/MKLau/Rclean/blob/b674396e4c84f58ae8a46ac50f7312650c300e07/R/package.R#L15) so they work out of the box, re https://github.com/ropensci/software-review/issues/327#issuecomment-562430994.
- [ ] `lintr::lint_package()` shows a bunch of new style issues. Some of them like whitespace issues are easy to fix. For others such as the snake case linter, you can use a top-level `.lintr` file to ignore them. Example: https://github.com/r-prof/proffer/blob/master/.lintr.

```
> packageVersion("lintr")
[1] ‘2.0.0’
> lintr::lint_package("Rclean")
...............
inst/example/long_script.R:25:33: style: There should be a space between right parenthesis and an opening curly brace.
for (i in seq_along(colnames(x))){
                                ^~
inst/example/long_script.R:58:1: style: Variable and function name style should be snake_case.
fit.23 <- lm(x2 ~ x3, data = data.frame(x2[, 1], x3[, 1]))
^~~~~~
inst/example/long_script.R:64:1: style: Variable and function name style should be snake_case.
fit.xx <- lm(A~B, data = x)
^~~~~~
inst/example/long_script.R:70:1: style: Variable and function name style should be snake_case.
fit.sqrt.A <- lm(I(sqrt(A))~B, data = x)
^~~~~~~~~~
inst/example/long_script.R:76:57: style: Trailing whitespace is superfluous.
## After that. I came back and ran another analysis with
                                                        ^
inst/example/long_script.R:79:25: style: Put spaces around all infix operators.
z <- c(rep("A", nrow(x2)/2), rep("B", nrow(x2)/2))
                       ~^~
inst/example/long_script.R:79:47: style: Put spaces around all infix operators.
z <- c(rep("A", nrow(x2)/2), rep("B", nrow(x2)/2))
                                             ~^~
inst/example/long_script.R:80:1: style: Variable and function name style should be snake_case.
fit.anova <- aov(x2 ~ z, data = data.frame(x2 = x2[, 1], z))
^~~~~~~~~
inst/example/micro.R:1:2: style: Put spaces around all infix operators.
x<- 1
~^
inst/example/micro.R:2:3: style: Put spaces around all infix operators.
y <-3
  ^~~

# There are more...
```
