### 2020-01-06 #327 @MKLau
Hi Will (@wlandau), I've now made the system.file() changes and fixed the lints in all the code files, and I've added an issues template for bug reporting. I went ahead and added lintr to Travis as an after-success check as well. One thing to note, I have not changed/fixed all of the lints (such as camel_case for all variables and spacing, etc.) for the example scripts, as those examples contain certain "issues" purposely to imitate quickly composed, "realistic" scripts.

All changes have been merged into the current master ([PR #195](https://github.com/MKLau/rclean)). Let me know your thoughts and if I've adequately addressed your last round of comments.

Cheers,

Matt
