### 2020-01-07 #327 @wlandau
Thanks, @MKLau. We're *almost* there. The very last issue on my end is some trouble running the updated example of `keep()`.

```r
script <- system.file(
        "example",
        "simple_script.R",
        package = "Rclean")
clean.code <- clean(script, “tab.15”)
#> Error: unexpected input in "clean.code <- clean(script, �"
```

I think it is because you are using `“` (probably unicode) and not `"` (ASCII 34) for quotes. Should be a simple fix [here](https://github.com/MKLau/Rclean/blob/1b430f0ca7590c8fad1308cddf8358b87bf6fe50/R/keep.R#L43).
