### 2020-01-07 #327 @MKLau
Fixed and committed to master.
