### 2020-01-07 #327 @wlandau
Confirmed, thanks.

You have addressed all my feedback, and  `Rclean` has come a long way in a short time. As a reviewer, I approve `Rclean` for rOpenSci. Well done, @MKLau!
