### 2020-01-08 #327 @annakrystalli
Thanks for your efforts all involved! 👏

I'm currently on the mountain 🏂 on my last day of hols but will set the wheels in motion for finalisation of approval tomorrow morning when I'm back at my laptop.

Sent from my iPhone

> On 8 Jan 2020, at 01:14, Will Landau <notifications@github.com> wrote:
>
> ﻿
> Confirmed, thanks.
>
> You have addressed all my feedback, and Rclean has come a long way in a short time. As a reviewer, I approve Rclean for rOpenSci. Well done, @MKLau!
>
