### 2020-01-09 #327 @danielskatz
Note that this is already under review by JOSS, and that [review](https://github.com/openjournals/joss-reviews/issues/1312) has been paused while the rOpenSci has proceeded.  Once this rOpenSci process is complete, the JOSS review can restart, and should be fast.
