### 2020-01-09 #327 @MKLau
Great, thanks Anna (@annakrystalli), will get on those ASAP.

Also, thanks @danielskatz for keeping the manuscript in a holding pattern. It has been greatly improved via the software review. Looking forward to getting it finished.
