### 2020-01-10 #327 @stefaniebutland
Congratulations @MKLau on passing review, and your pending JOSS publication. Would you be interested in writing a Tech Note for rOpenSci about your package? Anna suggests that it is an "interesting package that could really simplify creating reproducible examples and pairing down code to the essentials for producing a given output."

[Tech Notes](https://ropensci.org/technotes/) are written for an audience that wants details and should provide something a reader could not glean from the documentation itself.

If interested, after JOSS publication is probably best timing, so you could respond then. Instructions are here: https://github.com/ropensci/roweb2#contributing-a-blog-post. Typically you would submit a draft by pull request, I review, and we can publish within a week.
