### 2020-02-10 #327 @MKLau
Hi @annakrystalli, Almost done with the transfer to ROS and edits on the JOSS manuscript. Github has been throwing a slow Unicorn! message whenever I try and create a pull request now that Rclean has moved to ROpenSci. Will try again tomorrow to merge the new changes that have updates to Travis and Codecov.

A couple questions:

- There is a message at https://docs.ropensci.org/rodev/reference/add_ro_desc.html that says "No longer recommended!" Should I still add this per your instructions or is there a new method?
- Do you have a moment to look over the manuscript (see <joss branch> Rclean/joss/paper.Rmd)? If not no worries. I'll pass it along to the JOSS reviewers once I get it linked up to Zenodo correctly.

Thanks!

Matt
