### 2020-02-11 #327 @MKLau
Thanks for looking over the paper, @annakrystalli . I'll go ahead and add a line showing library("Rclean").

One more question, how do I enable Zenodo watching? I don't see it when I access Zenodo. Should I have done this prior to transferring the repo to ROpenSci?

Also, I keep getting a long loading time page.

![image](https://user-images.githubusercontent.com/5342857/74269680-d85eea00-4cd7-11ea-92a0-0175bc5305f8.png)

Do you have access to the repo? Would you be able to check if you can open a new pull request?
