### 2020-02-12 #327 @MKLau
@annakrystalli

- I'm still getting this slow loading message when I try to setup a pull request, so I just bypassed it and did a manual merge and push.
- Also, looks as though my previous Zenodo setup automatically switched from MKLau to ROpenSci. So, good to go.

I'm going to go ahead and send to JOSS. Let me know if you get a chance to look at the pull request initiation from your end though.
