### 2020-02-13 #327 @annakrystalli
Hi @MKLau, just looking into the pull request issues you're having. I've managed to make a successful PR (not merged). https://github.com/ropensci/Rclean/pull/201

When and where exactly are you getting the slow unicorn message?
