### 2020-02-13 #327 @annakrystalli
Good to hear it's all working now! I take it you are in the process of completing with JOSS too right? I'm going to go ahead and close this issue now.
