### 2020-02-26 #327 @stefaniebutland
Congratulations on your JOSS publication @MKLau. We would love to have a technote about Rclean as noted above
> Anna suggests that it is an "interesting package that could really simplify creating reproducible examples and pairing down code to the essentials for producing a given output.

We now have more detailed guidance: https://blogguide.ropensci.org/

If you're interested, please suggest a date for submission and I can provide a publication date.
