### 2020-03-03 #327 @stefaniebutland
Yes it would thank you. Please submit when you're ready using tentative publication date 2020-03-17.

cc @ropensci/blog-editors
