### 2020-03-13 #327 @MKLau
Hi @stefaniebutland,

Just finished with proof reading and spellcheck of the technote. You can find it here: https://github.com/ropensci/Rclean/blob/technote/ropensci/rclean_technote.md

The associated Rmd file is in the same directory on that branch too, if you need to have a look.

I’ve written it as an expanded version of the JOSS article, with a bit more discussion of a few details. Happy to have any edits and/or suggestions though.

Thanks and hope you’re healthy and well.
