### 2020-03-13 #327 @steffilazerte
Hi @MKLau, I've been chatting with @stefaniebutland and we were thinking that your article looks a bit more like a blog post than a technote (particularly the section talking about the Provenance Engine). This is great!

To get this published, I invite you to submit it as a pull request to the `roweb2` repository.
You can see the full instructions for setting it up in `roweb2` in the [Blog Guide](https://blogguide.ropensci.org/), particularly the chapter on [Technical Guidelines](https://blogguide.ropensci.org/technical.html). If you agree with our assessment of a blog post, rather than technote, just follow the set up for blog post.

I'll be your friendly reviewer and will be ready to review on Monday, if you can open the pull request by then. Once you're ready for review, either let me know in a comment on the pull request or change the pull request from Draft to Non-Draft.

Thanks!
